% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{getCommand.Psychometric}
\alias{getCommand.Psychometric}
\title{Summary of getCommandMissing}
\usage{
\method{getCommand}{Psychometric}(object, scale = "All", command = "Alpha")
}
\arguments{
\item{object}{A Reliability object}

\item{scale}{A Reliability object}

\item{command}{Alpha, Omega}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
