% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{getCommand.Reliability}
\alias{getCommand.Reliability}
\title{Makes it simple to do basic psychometrics}
\usage{
\method{getCommand}{Reliability}(object, scale = "All")
}
\arguments{
\item{object}{A Reliability object}

\item{scale}{All, Alpha, Omega}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
