% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{getObject.Psychometric}
\alias{getObject.Psychometric}
\title{Summary of getObject}
\usage{
\method{getObject}{Psychometric}(object)
}
\arguments{
\item{object}{A Reliability object}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
