% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPsychometric.R
\name{GetPsychometricJamovi}
\alias{GetPsychometricJamovi}
\title{Getting the Psychometric class}
\usage{
GetPsychometricJamovi(
  dataSetNr,
  scaleNames,
  responseScale = list(c(1, 5)),
  typeSum = "Mean",
  itemLength = 6,
  reverse = T,
  idVar = "ID"
)
}
\arguments{
\item{dataSetNr}{A dataframe with the items and eventual extra variables}

\item{scaleNames}{A vector with the names of the scales}

\item{responseScale}{A list with vectors of response intervalls, either one (if all are the same) or one for each scale}

\item{typeSum}{A functions used for summarizing the items to scales}

\item{itemLength}{The number of significant characters in items}

\item{reverse}{A boolean that sets whether items ending with 'R' should be reversed}

\item{idVar}{The name of the case identifier (if none a case number variable called ID will be created)}
}
\value{
A Psychometric object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
\examples{
myPsFrame <- GetPsychometricJamovi(myFrame, c("FirstScale", "SecondScale"), 
                    responseScale = list(c(1,7)))
}
