% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPsychometric.R
\name{imputeMissing.Psychometric}
\alias{imputeMissing.Psychometric}
\title{Makes it simple to do basic psychometrics}
\usage{
\method{imputeMissing}{Psychometric}(object, handleMissing = "Listwise", scales = F)
}
\arguments{
\item{object}{A Reliability object}

\item{handleMissing}{A Reliability object}

\item{scales}{A Reliability object}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
