% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Makes it simple to do basic psychometrics}
\usage{
imputeMissing(object, ...)
}
\arguments{
\item{object}{A Reliability object}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
