% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{plot.Reliability}
\alias{plot.Reliability}
\title{Summary of plot}
\usage{
\method{plot}{Reliability}(x, scale = NULL, ...)
}
\arguments{
\item{x}{A Reliability object}

\item{scale}{which scale}

\item{...}{Other arguments}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
