% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{plotScale.Psychometric}
\alias{plotScale.Psychometric}
\title{Summary of plotScale.Psychometric}
\usage{
\method{plotScale}{Psychometric}(
  object,
  scale = "All",
  group = NULL,
  external = NULL,
  type = "Histogram",
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{object}{A Reliability object}

\item{scale}{A Reliability object}

\item{group}{All, Alpha, Omega}

\item{external}{All, Alpha, Omega}

\item{type}{function}

\item{main}{function}

\item{xlab}{function}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
