% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPsychometric.R
\name{print.Reliability}
\alias{print.Reliability}
\title{Summary of print}
\usage{
\method{print}{Reliability}(object, plot = F)
}
\arguments{
\item{object}{A Reliability object}

\item{plot}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
