% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{reliabilityTest.Psychometric}
\alias{reliabilityTest.Psychometric}
\title{Getting the Psychometric class}
\usage{
\method{reliabilityTest}{Psychometric}(object, what = "Alpha", ...)
}
\arguments{
\item{object}{A Psychometric object}

\item{what}{what type of analyses: Alpha, Omega or Parallel}

\item{...}{gh}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
