% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{split.Psychometric}
\alias{split.Psychometric}
\title{Summary of split}
\usage{
\method{split}{Psychometric}(object, group, f, ...)
}
\arguments{
\item{object}{A Reliability object}

\item{group}{All, Alpha, Omega}

\item{f}{function}

\item{...}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
