% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPsychometric.R
\name{summary.Reliability}
\alias{summary.Reliability}
\title{Summary of Reliability}
\usage{
\method{summary}{Reliability}(object, scale = NULL)
}
\arguments{
\item{object}{A Reliability object}

\item{scale}{which scale}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
\examples{
myPsFrame <- summary(object)
}
