% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestFunctions.R
\name{print.Reliability}
\alias{print.Reliability}
\title{Summary of print}
\usage{
\method{print}{Reliability}(x, plot = F, ...)
}
\arguments{
\item{x}{A Reliability object}

\item{plot}{which scale}

\item{...}{Other arguments}
}
\value{
A Reliability object that can be used for analyses
}
\description{
Makes it simple to do basic psychometrics
}
