% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPsychometric.R
\name{print.Reliability}
\alias{print.Reliability}
\title{print the Reliability Object}
\usage{
\method{print}{Reliability}(object, plot = F)
}
\arguments{
\item{object}{Reliability Object}

\item{plot}{kind of print}
}
\value{

}
\description{
print the Reliability Object
}
\examples{
print(object)
}
